Rework file structure
Make more explicit what are packages and what are binaries to be
compiled.

Package utilities inside pkg
Group binaries inside cmd

Provide a backward compatible `main.go` entrypoint
