+++ pkg/kubectl/test-home/.kube/bin/kubectl-darwin-amd64-0.0.0.2
@@ -1,0 +1,3 @@
+#!/bin/bash
+echo $@
+exit 0
